% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstalk.R
\name{crosstalkLibs}
\alias{crosstalkLibs}
\title{Crosstalk dependencies}
\usage{
crosstalkLibs()
}
\description{
List of \code{\link[htmltools]{htmlDependency}} objects necessary for
Crosstalk to function. Intended for widget authors.
}
