% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{de_res}
\alias{de_res}
\title{Example differential expression results}
\format{
A data frame containing gene-level differential expression statistics.
}
\usage{
data(de_res)
}
\description{
Example results of differential expression analysis used to demonstrate gene
prioritization functionality in CrosstalkX.
}
\keyword{datasets}
