% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-shims.R
\name{getDefaultReactiveDomain}
\alias{getDefaultReactiveDomain}
\title{Get default reactive domain}
\usage{
getDefaultReactiveDomain()
}
\description{
Pass-through to [shiny::getDefaultReactiveDomain()], unless the shiny package
is not installed, in which case `NULL` is returned.
}
\keyword{internal}
