% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstalk.R
\name{is.SharedData}
\alias{is.SharedData}
\title{Check if an object is \code{SharedData}}
\usage{
is.SharedData(x)
}
\arguments{
\item{x}{The object that may or may not be an instance of \code{SharedData}}
}
\value{
logical
}
\description{
Check if an object is an instance of \code{\link{SharedData}} or not.
}
