% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{maintain_selection}
\alias{maintain_selection}
\title{Synchronize Shiny brush selection with shared data}
\usage{
maintain_selection(sharedData, brushId, ownerId = "")
}
\arguments{
\item{sharedData}{The shared data instance}

\item{brushId}{Character vector indicating the name of the \code{plotOutput}
brush}

\item{ownerId}{(TBD)}
}
\description{
Waits for a brush to change, and propagates that change to the
\code{sharedData} object.
}

